+++ zipline/data/_adjustments.pyx
