+++ zipline/data/_equities.pyx
